requirements.txt
asyncpg==0.29.0
beautifulsoup4==4.12.3
branca==0.7.2
certifi==2024.6.2
charset-normalizer==3.3.2
et-xmlfile==1.1.0
folium==0.16.0
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
openpyxl==3.1.4
packaging==24.1
pandas==2.2.2
playwright==1.44.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pyee==11.1.0
pytest==8.2.2
pytest-asyncio==0.23.7
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
shapely==2.0.4
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.30
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
xyzservices==2024.6.0
